chapter_004/src/test/java/ru/job4j/collectionpro/IteratorForSimpleNumberTest.java
package ru.job4j.collectionpro;

import org.junit.Test;
import java.util.NoSuchElementException;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;


public class IteratorForSimpleNumberTest {

    @Test
    public void shouldReturnPrimeNumbersOnly() {
        IteratorForSimpleNumber it = new IteratorForSimpleNumber();
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(2));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(3));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(5));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(7));
        assertThat(it.hasNext(), is(true));
        assertThat(it.next(), is(3571));
    }

}